requirements.txt
Django==5.2.8
djangorestframework
whitenoise
gunicorn
